feat: Add categories to layout components
This commit adjusts layout components to handle categories. In particular, `categories` prop was added to the `LayoutHeader`, `LayoutNavButton`, `LayoutNavDrawer`, and `LayoutNav`. Modifications were also made to the structure of `LayoutNav` to better utilize the added `categories`. This change broadens the flexibility of the layout components allowing them to dynamically adapt based on the categories given. Additional changes were made to mechanism used within some pages to provide the categories.